requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-decouple==3.8
psycopg==3.1.16
celery==5.3.6
redis==5.0.1
qrcode==7.4.2
reportlab==4.0.9
django-filter==23.5
drf-spectacular==0.27.1
cryptography==42.0.2
uuid==1.30
django-extensions==3.2.3
django-debug-toolbar==4.2.0
whitenoise==6.6.0
gunicorn==21.2.0
